fix: notch appears when the mouse cursor is over the mini-overlay
